migrated individual client command parsers to PersonCommandParser class
