+++ src/main/java/seedu/address/logic/commands/person/AddPersonCommand.java
@@ -16,7 +16,7 @@ import seedu.address.model.person.Person;
 /**
  * Adds a person to the address book.
  */
-public class AddPersonCommand extends Command {
+public class AddPersonCommand extends PersonCommand {
 
     public static final String COMMAND_WORD = "add";
