+++ src/main/java/seedu/address/logic/commands/person/EditPersonCommand.java
@@ -31,7 +31,7 @@ import seedu.address.model.tag.Tag;
 /**
  * Edits the details of an existing person in the address book.
  */
-public class EditPersonCommand extends Command {
+public class EditPersonCommand extends PersonCommand {
 
     public static final String COMMAND_WORD = "edit";
