+++ src/main/java/seedu/address/logic/commands/person/PersonCommand.java
@@ -3,5 +3,5 @@ package seedu.address.logic.commands.person;
 import seedu.address.logic.commands.Command;
 
 public abstract class PersonCommand extends Command {
-    public static final String COMMAND_STR = "client";
+    public static final String COMMAND_WORD = "client";
 }
