+++ src/main/java/seedu/address/logic/commands/person/TagPersonCommand.java
@@ -1,0 +1,16 @@
+package seedu.address.logic.commands.person;
+
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
+import seedu.address.model.person.Person;
+
+/**
+ * TODO: Implement this to tag a person to a project
+ */
+public class TagPersonCommand extends PersonCommand {
+    @Override
+    public CommandResult execute(Model model) throws CommandException {
+        return null;
+    }
+}
