+++ src/main/java/seedu/address/logic/parser/FindCommandParser.java
@@ -30,4 +30,9 @@ public class FindCommandParser implements Parser<FindCommand> {
         return new FindCommand(new NameContainsKeywordsPredicate(Arrays.asList(nameKeywords)));
     }
 
+    // TODO
+    @Override
+    public FindCommand parse(String flag, String arguments) throws ParseException {
+        return null;
+    }
 }
