+++ src/main/java/seedu/address/logic/parser/Parser.java
@@ -12,5 +12,6 @@ public interface Parser<T extends Command> {
      * Parses {@code userInput} into a command and returns it.
      * @throws ParseException if {@code userInput} does not conform the expected format
      */
-    T parse(String userInput) throws ParseException;
+//    T parse(String userInput) throws ParseException;
+    T parse(String flag, String arguments)throws ParseException;
 }
