+++ src/main/java/seedu/address/logic/parser/exceptions/ParseException.java
@@ -1,6 +1,6 @@
 package seedu.address.logic.parser.exceptions;
 
-import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.model.tag.exceptions.IllegalValueException;
 
 /**
  * Represents a parse error encountered by a parser.
