+++ src/main/java/seedu/address/model/tag/exceptions/DataConversionException.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.exceptions;
+package seedu.address.model.tag.exceptions;
 
 /**
  * Represents an error during conversion of data from one format to another
