+++ src/main/java/seedu/address/model/tag/exceptions/IllegalValueException.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.exceptions;
+package seedu.address.model.tag.exceptions;
 
 /**
  * Signals that some given data does not fulfill some constraints.
