+++ src/main/java/seedu/address/storage/JsonAddressBookStorage.java
@@ -8,8 +8,8 @@ import java.util.Optional;
 import java.util.logging.Logger;
 
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.commons.exceptions.DataConversionException;
-import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.model.tag.exceptions.DataConversionException;
+import seedu.address.model.tag.exceptions.IllegalValueException;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.commons.util.JsonUtil;
 import seedu.address.model.ReadOnlyAddressBook;
